+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 
 ### v0.2.0 - unreleased
-
+ - **FEATURE** Introduce a public `teardown()` method for on-the-fly instances
 
 ### v0.1.0 - September 20th, 2020
  - Documentation re-written